Bugfixes for vim 7.4
Fixed bug where Conque tried to unmap keys that weren't mapped in
the first place if an application exited normally. The solution is
a hack: I disabled all attempts to unmap keys. They're all
buffer-local, so it doesn't matter; the buffer is being destroyed
anyway, and that clears them all.

Fixed bug where Conque allowed you to type characters even after
the application exited normally. Now, it correctly sets the buffer
to nomodifiable.

Fixed bug where Conque stayed in insert mode when an application
exited normally. It was only kicking the user out of insert mode
if the option to destroy the buffer immediately upon closing was
set. Now it kicks the user out, and THEN checks for that option.